Crank Osu puzzle
Implemented the crank Osu-style puzzle. Created an instance of
CrankOsu object in Scene. One can set how fast the notes appear
with the tempo argument which specifies the amount of delay
between note spawns. The winScore argument specifies the number
of notes to be successfully let through for the puzzle to be
completed.